requirements.txt
asgiref==3.2.10
certifi==2020.12.5
dj-database-url==0.5.0
django==3.1.2
django-cors-headers==3.4.0
django-environ==0.4.5
djangorestframework==3.12.2
PyJWT==1.7.1
djangorestframework_simplejwt==4.3.0
gunicorn==20.0.4
olefile==0.46
pillow==8.0.0
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
